refactor(codex): split goal routes and parity tests
