+++ omnigent/server/routes/codex/__init__.py
@@ -1,0 +1,1 @@
+"""Codex-specific route modules."""
