+++ tests/codex_parity/__init__.py
@@ -1,0 +1,1 @@
+"""Codex parity test package."""
